requirements.txt
pandas~=1.2.3
omegaconf~=2.1.1
torchvision~=0.12.0
pillow~=9.0.1
sqlalchemy~=1.4.35
sklearn~=0.0
scikit-learn~=1.1.1
numpy~=1.21.2
transformers~=4.17.0
gensim~=4.1.2
datasets~=2.3.3.dev0